+++ lib/rspec/mocks.rb
@@ -26,6 +26,10 @@ module RSpec
         space.proxy_for(object)
       end
 
+      def proxies_of(klass)
+        space.proxies_of(klass)
+      end
+
       def any_instance_recorder_for(klass)
         space.any_instance_recorder_for(klass)
       end
@@ -43,7 +47,9 @@ module RSpec
       #   x = 0
       #   RSpec::Mocks.allow_message(bar, :foo) { x += 1 }
       def allow_message(subject, message, opts={}, &block)
-        orig_caller = opts.fetch(:expected_from) { caller(1)[0] }
+        orig_caller = opts.fetch(:expected_from) {
+          CallerFilter.first_non_rspec_line
+        }
         ::RSpec::Mocks.proxy_for(subject).
           add_stub(orig_caller, message.to_sym, opts, &block)
       end
@@ -60,7 +66,9 @@ module RSpec
       #   RSpec::Mocks.expect_message(bar, :foo)
       #   bar.foo
       def expect_message(subject, message, opts={}, &block)
-        orig_caller = opts.fetch(:expected_from) { caller(1)[0] }
+        orig_caller = opts.fetch(:expected_from) {
+          CallerFilter.first_non_rspec_line
+        }
         ::RSpec::Mocks.proxy_for(subject).
           add_message_expectation(orig_caller, message.to_sym, opts, &block)
       end