+++ lib/rspec/mocks/any_instance/recorder.rb
@@ -11,10 +11,11 @@ module RSpec
       # @see Chain
       class Recorder
         # @private
-        attr_reader :message_chains
+        attr_reader :message_chains, :stubs
 
         def initialize(klass)
           @message_chains = MessageChains.new
+          @stubs = Hash.new { |hash,key| hash[key] = [] }
           @observed_methods = []
           @played_methods = {}
           @klass = klass
@@ -26,13 +27,13 @@ module RSpec
         #
         # @see Methods#stub
         def stub(method_name_or_method_map, &block)
-          if method_name_or_method_map.is_a?(Hash)
+          if Hash === method_name_or_method_map
             method_name_or_method_map.each do |method_name, return_value|
               stub(method_name).and_return(return_value)
             end
           else
             observe!(method_name_or_method_map)
-            message_chains.add(method_name_or_method_map, StubChain.new(method_name_or_method_map, &block))
+            message_chains.add(method_name_or_method_map, StubChain.new(self, method_name_or_method_map, &block))
           end
         end
 
@@ -44,7 +45,7 @@ module RSpec
         def stub_chain(*method_names_and_optional_return_values, &block)
           normalize_chain(*method_names_and_optional_return_values) do |method_name, args|
             observe!(method_name)
-            message_chains.add(method_name, StubChainChain.new(*args, &block))
+            message_chains.add(method_name, StubChainChain.new(self, *args, &block))
           end
         end
 
@@ -56,7 +57,7 @@ module RSpec
         def should_receive(method_name, &block)
           @expectation_set = true
           observe!(method_name)
-          message_chains.add(method_name, PositiveExpectationChain.new(method_name, &block))
+          message_chains.add(method_name, PositiveExpectationChain.new(self, method_name, &block))
         end
 
         def should_not_receive(method_name, &block)
@@ -72,6 +73,10 @@ module RSpec
             raise RSpec::Mocks::MockExpectationError, "The method `#{method_name}` was not stubbed or was already unstubbed"
           end
           message_chains.remove_stub_chains_for!(method_name)
+          ::RSpec::Mocks.proxies_of(@klass).each do |proxy|
+            stubs[method_name].each { |stub| proxy.remove_single_stub(method_name, stub) }
+          end
+          stubs[method_name].clear
           stop_observing!(method_name) unless message_chains.has_expectation?(method_name)
         end