+++ lib/rspec/mocks/test_double.rb
@@ -68,7 +68,7 @@ module RSpec
     private
 
       def __initialize_as_test_double(name=nil, stubs={})
-        if name.is_a?(Hash) && stubs.empty?
+        if Hash === name && stubs.empty?
           stubs = name
           @name = nil
         else
@@ -97,7 +97,7 @@ module RSpec
 
       def assign_stubs(stubs)
         stubs.each_pair do |message, response|
-          Mocks.allow_message(self, message).and_return(response)
+          __mock_proxy.add_simple_stub(message, response)
         end
       end