+++ lib/rspec/mocks/verifying_proxy.rb
@@ -31,6 +31,11 @@ module RSpec
         super
       end
 
+      def add_simple_stub(method_name, *args)
+        ensure_implemented(method_name)
+        super
+      end
+
       def add_message_expectation(location, method_name, opts={}, &block)
         ensure_implemented(method_name)
         super
@@ -52,8 +57,6 @@ module RSpec
           h[k] = method_double
         end
       end
-
-    protected
 
       def ensure_implemented(method_name)
         @doubled_module.when_loaded do |original_module|