+++ spec/rspec/mocks/any_instance/message_chains_spec.rb
@@ -1,9 +1,10 @@
 require 'spec_helper'
 
 describe RSpec::Mocks::AnyInstance::MessageChains do
+  let(:recorder) { double }
   let(:chains) { RSpec::Mocks::AnyInstance::MessageChains.new }
-  let(:stub_chain) { RSpec::Mocks::AnyInstance::StubChain.new }
-  let(:expectation_chain) { RSpec::Mocks::AnyInstance::PositiveExpectationChain.new }
+  let(:stub_chain) { RSpec::Mocks::AnyInstance::StubChain.new recorder }
+  let(:expectation_chain) { RSpec::Mocks::AnyInstance::PositiveExpectationChain.new recorder }
 
   it "knows if a method does not have an expectation set on it" do
     chains.add(:method_name, stub_chain)
@@ -19,7 +20,7 @@ describe RSpec::Mocks::AnyInstance::MessageChains do
   it "can remove all stub chains" do
     chains.add(:method_name, stub_chain)
     chains.add(:method_name, expectation_chain)
-    chains.add(:method_name, RSpec::Mocks::AnyInstance::StubChain.new)
+    chains.add(:method_name, RSpec::Mocks::AnyInstance::StubChain.new(recorder))
 
     chains.remove_stub_chains_for!(:method_name)
     expect(chains[:method_name]).to eq([expectation_chain])
@@ -33,7 +34,7 @@ describe RSpec::Mocks::AnyInstance::MessageChains do
 
     it "allows multiple stub chains for a method" do
       chains.add(:method_name, stub_chain)
-      chains.add(:method_name, another_stub_chain = RSpec::Mocks::AnyInstance::StubChain.new)
+      chains.add(:method_name, another_stub_chain = RSpec::Mocks::AnyInstance::StubChain.new(recorder))
       expect(chains[:method_name]).to eq([stub_chain, another_stub_chain])
     end
   end