+++ spec/rspec/mocks/double_spec.rb
@@ -14,4 +14,11 @@ describe "double" do
     m = double('cup')
     expect(m.inspect).to match(/#<RSpec::Mocks::Mock:0x[a-f0-9.]+ @name="cup">/)
   end
+
+  it 'restores standard object methods on reset' do
+    dbl = double(:tainted? => true)
+    expect(dbl.tainted?).to eq(true)
+    reset dbl
+    expect(dbl.tainted?).to eq(false)
+  end
 end