+++ spec/spec_helper.rb
@@ -30,12 +30,19 @@ module VerifyAndResetHelpers
   end
 end
 
+module DeprecationHelpers
+  def expect_deprecation_with_call_site(file, line)
+    expect(RSpec.configuration.reporter).to receive(:deprecation) do |options|
+      expect(options[:call_site]).to include([file, line].join(':'))
+    end
+  end
+end
+
 RSpec.configure do |config|
   config.mock_with :rspec
   config.color_enabled = true
   config.order = :random
   config.run_all_when_everything_filtered = true
-  config.treat_symbols_as_metadata_keys_with_true_values = true
   config.filter_run_including :focus
 
   config.expect_with :rspec do |expectations|
@@ -53,6 +60,7 @@ RSpec.configure do |config|
   end
 
   config.include VerifyAndResetHelpers
+  config.include DeprecationHelpers
 end
 
 shared_context "with syntax" do |syntax|
@@ -80,3 +88,4 @@ shared_context "with isolated configuration" do
     RSpec::Mocks.instance_variable_set(:@configuration, orig_configuration)
   end
 end
+